% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterKernelEstimates.R
\docType{package}
\name{rasterKernelEstimates}
\alias{rasterKernelEstimates}
\alias{rasterKernelEstimates-package}
\title{rasterKernelEstimates:  Kernel Based Estimates on In-Memory Raster Images}
\description{
rasterKernelEstimats provides kernel based estimates for in-memory raster images 
from the raster package.  These kernel estimates include local means
variances, modes, medians, and Gaussian kernel density estimates.  All
results are in the form of raster images, preserving original resolution
and projection attributes in the raster image.
}
\details{
It has two main goals:

\itemize{
\item Provide a method to quickly perform common actions on in-memory raster
  images. 
\item Provide fast performance for in-memory raster images through C and OpenMP.
}
}

