% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterLocalCategoricalModes.R
\name{rasterLocalCategoricalModes}
\alias{rasterLocalCategoricalModes}
\title{Local categorical modes for an in memory raster image.}
\usage{
rasterLocalCategoricalModes(r, W)
}
\arguments{
\item{r}{An in memory raster image. Pixels should be whole numbers or \code{NA}. 
Pixels with non whole number values will be coerced to whole numbers.}

\item{W}{A matrix of weights.  The modal kernel will be applied at each 
centroid.  Dimensions must be non-zero and odd.}
}
\value{
An in memory raster image by most popular categories.
}
\description{
\code{rasterLocalCategoricalModes} finds the most popular category within the 
weighted neighborhood of W.
}
\examples{
r <- raster::raster( matrix(runif(81),9,9)) 
W <- matrix(1,3,3)
modeR <- rasterLocalCategoricalModes(r,W)
}

