% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterLocalMedians.R
\name{rasterLocalMedians}
\alias{rasterLocalMedians}
\title{Local medians for an in memory raster image.}
\usage{
rasterLocalMedians(r, W)
}
\arguments{
\item{r}{An in memory raster image.}

\item{W}{A matrix of weights.  The median kernel will be applied at each 
centroid.  Dimensions must be non-zero and odd.  Only non-missing or 
neighbors with non-zero weights are used in the median calculation.}
}
\value{
An in memory raster image of local medians.
}
\description{
\code{rasterLocalMedians} finds the median within the positive valued neighborhood 
of W.
}
\examples{
r <- raster::raster( matrix(rnorm(36),6,6)) 
W <- matrix(1,3,3)
medianR <- rasterLocalMedians(r,W)
}

