% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSpatialKDE.R
\name{rasterSpatialKDE}
\alias{rasterSpatialKDE}
\title{Calculates a local guassian kernel estimate for an in memory raster image.}
\usage{
rasterSpatialKDE(r, h, h.x, h.y)
}
\arguments{
\item{r}{An in memory raster image.}

\item{h}{the bandwidth parameter for the raster values.}

\item{h.x}{the bandwidth parameter for the raster columns.}

\item{h.y}{the bandwidth parameter for the raster rows.}
}
\value{
A raster image of the kernel density estimate.
}
\description{
\code{rasterLocalMoments} calculates a kernel density estimate truncated by 
the weighted neighborhood of W.
}
\examples{
r <- raster::raster( matrix(rnorm(36),6,6)) 
W <- matrix(1,3,3)
rLocalKDE <- rasterSpatialKDE(r,W)
}

