% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterLocalCounts.R
\name{rasterLocalSums}
\alias{rasterLocalSums}
\title{Local sums for an in memory raster image.}
\usage{
rasterLocalSums(r, W)
}
\arguments{
\item{r}{An in memory raster image.}

\item{W}{A matrix of weights.  The sums will be applied at each centroid. 
Dimensions must be non-zero and odd.  Only non-missing neighbors are used in 
the sum.}
}
\value{
An in memory raster image of local sums.
}
\description{
\code{rasterLocalSums} finds the local sum within the weighted neighborhood of W.
}
\examples{
r <- raster::raster( matrix(rnorm(36),6,6)) 
W <- matrix(1,3,3)
sumR <- rasterLocalSums(r,W)
}

